GOF-21 Updated add employee and trucks UI
Added UI to delete and update trucks

Changed look of add employee UI to mirror add trucks UI
